moving a method to a file
